+++ Simple.BTC-RPC/Simple.BTC/Models/Blcokchain/GetChainTips.cs
@@ -2,9 +2,9 @@
 
 public class GetChainTips_Result
 {
-    public int height { get; set; }
+    public long height { get; set; }
     public string hash { get; set; }
-    public int branchlen { get; set; }
+    public long branchlen { get; set; }
     public string status { get; set; }
 }
